Conflicts:
	src/condor_collector.V6/collector_engine.cpp
	src/condor_io/condor_ipverify.cpp
	src/condor_io/sock.cpp
	src/condor_schedd.V6/schedd.cpp
	src/condor_startd.V6/startd_main.cpp
	src/condor_syscall_lib/special_stubs.cpp
	src/condor_utils/cbstp_url.cpp
	src/condor_utils/ftp_url.cpp
	src/condor_utils/http_url.cpp
	src/condor_utils/my_hostname.h

+++ src/condor_utils/env.cpp
@@ -19,7 +19,7 @@
 
 
 #include "condor_common.h"
-#include "condor_attrlist.h"
+#include "condor_classad.h"
 #include "condor_string.h"
 #include "condor_attributes.h"
 
@@ -309,7 +309,7 @@ Env::WriteToDelimitedString(char const *input,MyString &output) {
 
 	while(*input) {
 		end = input + strcspn(input,specials);
-		ret = output.sprintf_cat("%.*s",end-input,input);
+		ret = output.sprintf_cat("%.*s", (int)(end-input), input);
 		ASSERT(ret);
 		input = end;
